+++ components/card/types/images.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from "react";
+
+import { Blob } from "../../../src/api/query";
+
+export interface Properties {
+  blob: Blob;
+  onLoad?: () => void;
+}
+
+export const ImageCard: FC<Properties> = ({ blob, onLoad }) => {
+  return <img src={blob.url} onLoad={onLoad} />;
+};
